# docs/openapi.yaml
openapi: 3.0.0
info:
  title: InnoAI API
  version: 3.1.0
  description: |
    AI Models for Every Filipino
    
    The InnoAI platform provides access to cutting-edge AI models including chat, image generation, 
    audio processing, 3D generation, video generation, and embeddings. Built for the Philippine market 
    with local currency pricing (PHP) and payment integration.
    
    ## Authentication
    
    The API supports three authentication methods:
    
    1. **JWT Bearer Token** - For customer portal authentication
    2. **API Key** - For inference endpoints (X-API-Key header)
    3. **Admin Key** - For administrative endpoints (x-admin-key header)
    
    ## Access Control Levels
    
    | Tier | Price | Requests/Day | Tokens/Month | Concurrent | Models |
    |------|-------|--------------|--------------|------------|--------|
    | Free | ₱0 | 100 | 10K | 1 | Basic only |
    | Starter | ₱499 | 1,000 | 100K | 3 | All chat |
    | Professional | ₱1,499 | 10,000 | 1M | 10 | All models |
    | Enterprise | ₱4,999 | 100,000 | 10M | 50 | All + Priority |
    
    ## Pricing (PHP)
    
    - **Chat (Boyong Opus)**: ₱465/1M input, ₱2,325/1M output
    - **Images (FLUX Pro)**: ₱3.75/image
    - **Video (Wan 480p)**: ₱4.68/second
    - **3D Generation**: Starting at ₱0.50/model
    
    ## Base URL
    
    All API requests should be made to: `https://ai-gateway.innoserver.cloud`
    
  contact:
    name: InnovateHub Support
    email: support@innovatehub.ph
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
  - AdminKeyAuth: []
...
paths:
...
  /api/customer/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new customer
      description: Create a new customer account with email and password
      operationId: registerCustomer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: customer@example.com
              password: SecurePass123!
              name: John Doe
              company: ACME Corp
      responses:
        '201':
          description: Customer registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
              example:
                success: true
                customer:
                  id: cust_abc123def456
                  email: customer@example.com
                  name: John Doe
                  company: ACME Corp
                  tier: free
                  status: active
                  createdAt: "2026-02-21T10:30:00.000Z"
                message: Registration successful. Please check your email to verify.
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/customer/auth/login:
    post:
      tags:
        - Authentication
      summary: Customer login
      description: Authenticate with email and password to receive JWT token
      operationId: loginCustomer
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: customer@example.com
              password: SecurePass123!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              example:
                success: true
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                customer:
                  id: cust_abc123def456
                  email: customer@example.com
                  name: John Doe
                  tier: professional
                  status: active
                expiresIn: 7d
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/customer/auth/profile:
    get:
      tags:
        - Authentication
      summary: Get customer profile
      description: Retrieve current customer's profile information
      operationId: getProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
              example:
                success: true
                customer:
                  id: cust_abc123def456
                  email: customer@example.com
                  name: John Doe
                  company: ACME Corp
                  tier: professional
                  status: active
                  createdAt: "2026-02-21T10:30:00.000Z"
                  lastLogin: "2026-02-21T15:45:00.000Z"
                  loginCount: 42
                  settings:
                    notifications: true
                    newsletter: true
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Authentication
      summary: Update customer profile
      description: Update customer profile information (name, company, settings)
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
            example:
              name: Jane Doe
              company: New Company Inc
              settings:
                notifications: true
                newsletter: false
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/customer/auth/change-password:
    post:
      tags:
        - Authentication
      summary: Change password
      description: Change customer password (requires current password)
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: OldPass123!
              newPassword: NewSecurePass456!
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password updated successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== CUSTOMER ENDPOINTS ====================
...
  /admin/billing/config:
    get:
      tags:
        - Billing
      summary: Get DirectPay configuration
      description: Get current DirectPay payment gateway configuration (Admin only)
      operationId: getBillingConfig
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Configuration retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  config:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/billing/environment:
    post:
      tags:
        - Billing
      summary: Set environment
      description: Set DirectPay environment to sandbox or production (Admin only)
      operationId: setBillingEnvironment
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                environment:
                  type: string
                  enum: [sandbox, production]
            example:
              environment: production
      responses:
        '200':
          description: Environment set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  config:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/billing/credentials:
    post:
      tags:
        - Billing
      summary: Save production credentials
      description: Save DirectPay production credentials (Admin only)
      operationId: saveBillingCredentials
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                merchantId:
                  type: string
                merchantKey:
                  type: string
                apiBase:
                  type: string
                dashboard:
                  type: string
                username:
                  type: string
                password:
                  type: string
            example:
              merchantId: MERCHANT123
              merchantKey: secret_key_abc123
      responses:
        '200':
          description: Credentials saved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/billing/test:
    post:
      tags:
        - Billing
      summary: Test DirectPay connection
      description: Test the DirectPay payment gateway connection (Admin only)
      operationId: testBillingConnection
      security:
        - AdminKeyAuth: []
      responses:
        '200':
          description: Connection test result
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/billing/payment:
    post:
      tags:
        - Billing
      summary: Create payment
      description: Create a payment/checkout session (Admin only)
      operationId: createPayment
      security:
        - AdminKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                description:
                  type: string
                metadata:
                  type: object
            example:
              amount: 1499
              description: Professional Plan Subscription
              metadata:
                customerId: cust_abc123
      responses:
        '200':
          description: Payment created successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/billing/transactions:
    get:
      tags:
        - Billing
      summary: Get transactions
      description: Get list of payment transactions (Admin only)
      operationId: getTransactions
      security:
        - AdminKeyAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Transactions retrieved successfully
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== INFERENCE ENDPOINTS ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for customer authentication
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for inference endpoints (format ik_xxxxxxxx...)
    AdminKeyAuth:
      type: apiKey
      in: header
      name: x-admin-key
      description: Admin key for administrative endpoints
  schemas:
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - name
      properties:
        email:
          type: string
          format: email
          example: customer@example.com
        password:
          type: string
          minLength: 8
          example: SecurePass123!
        name:
          type: string
          example: John Doe
        company:
          type: string
          example: ACME Corp
        tier:
          type: string
          enum: [free, starter, professional, enterprise]
          default: free
    RegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        customer:
          $ref: '#/components/schemas/Customer'
        message:
          type: string
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        token:
          type: string
          description: JWT token for authentication
        customer:
          $ref: '#/components/schemas/Customer'
        expiresIn:
          type: string
          example: 7d
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        company:
          type: string
        settings:
          type: object
          properties:
            notifications:
              type: boolean
            newsletter:
              type: boolean
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8
    ProfileResponse:
      type: object
      properties:
        success:
          type: boolean
        customer:
          $ref: '#/components/schemas/Customer'

    # ==================== CUSTOMER SCHEMAS ====================
    Customer:
      type: object
      properties:
        id:
          type: string
          example: cust_abc123def456
        email:
          type: string
          format: email
        name:
          type: string
        company:
          type: string
        tier:
          type: string
          enum: [free, starter, professional, enterprise]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
        lastLogin:
          type: string
          format: date-time
          nullable: true
        loginCount:
          type: integer
        settings:
          type: object
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Invalid request parameters
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
          example:
            error: Authentication required
...